refactor(account): update layout and menu structure for improved navigation
