+++ frontend/src/app/(account)/account.module.css
@@ -5,7 +5,7 @@
   border-bottom: 1px solid var(--color-home-grey);
 }
 
-.buttonContainer {
+.buttonsContainer {
   display: flex;
   flex-direction: column;
   gap: 14px;
