+++ frontend/src/app/(account)/studentMenu.ts
@@ -1,8 +1,8 @@
 export const studentMenu = [
-  { href: 'settings', label: 'Настройки профиля' },
-  { href: 'applications', label: 'Мои заявки на программы' },
-  { href: 'events', label: 'Мои записи на мероприятия' },
-  { href: 'faq', label: 'Вопросы и ответы' },
-  { href: 'contacts', label: 'Контакты и адрес' },
-  { href: 'license', label: 'Лицензия и аккредитация' },
+  { href: 'student/settings', label: 'Настройки профиля' },
+  { href: 'student/applications', label: 'Мои заявки на программы' },
+  { href: 'student/events', label: 'Мои записи на мероприятия' },
+  { href: 'student/faq', label: 'Вопросы и ответы' },
+  { href: 'student/contacts', label: 'Контакты и адрес' },
+  { href: 'student/license', label: 'Лицензия и аккредитация' },
 ];
